+++ tests/lib/js/static_interop_test/external_member_type_parameters_static_test.dart
@@ -22,6 +22,8 @@ external T invalidTopLevel<T>(T t);
 
 typedef Typedef<T> = T Function();
 
+extension type JSList<T>._(JSAny? _) {}
+
 @JS()
 @staticInterop
 class Uninstantiated<W, X extends Instantiated?> {
@@ -68,12 +70,11 @@ extension UninstantiatedExtension<T, U extends JSAny?, V extends Instantiated>
   external void consumeV(V v);
 
   // Test type parameters in a nested type context.
-  Set<Typedef<T>> get getNestedTDart => throw UnimplementedError();
-  external Set<Typedef<T>> get getNestedT;
-  //                           ^
-  // [web] External static interop members can only use type parameters that extend either a static interop type or one of the 'dart:js_interop' types.
-  external Set<Typedef<U>> get getNestedU;
-  external Set<Typedef<V>> get getNestedV;
+  JSList<Typedef<T>> get getNestedTDart => throw UnimplementedError();
+  // No error as JSList is an interop extension type.
+  external JSList<Typedef<T>> get getNestedT;
+  external JSList<Typedef<U>> get getNestedU;
+  external JSList<Typedef<V>> get getNestedV;
 
   // Test type parameters that are declared by the member.
   W returnWDart<W>() => throw UnimplementedError();
@@ -126,12 +127,11 @@ extension type UninstantiatedExtensionType<T, U extends JSAny?,
   external void consumeV(V v);
 
   // Test type parameters in a nested type context.
-  Set<Typedef<T>> get getNestedTDart => throw UnimplementedError();
-  external Set<Typedef<T>> get getNestedT;
-  //                           ^
-  // [web] External static interop members can only use type parameters that extend either a static interop type or one of the 'dart:js_interop' types.
-  external Set<Typedef<U>> get getNestedU;
-  external Set<Typedef<V>> get getNestedV;
+  JSList<Typedef<T>> get getNestedTDart => throw UnimplementedError();
+  // No error as JSList is an interop extension type.
+  external JSList<Typedef<T>> get getNestedT;
+  external JSList<Typedef<U>> get getNestedU;
+  external JSList<Typedef<V>> get getNestedV;
 
   // Test type parameters that are declared by the member.
   W returnWDart<W>() => throw UnimplementedError();
@@ -178,25 +178,25 @@ extension PkgJsExtension<T> on PkgJs<T> {
 @JS()
 @staticInterop
 class Instantiated {
-  external factory Instantiated(List<JSNumber> list);
+  external factory Instantiated(JSList<JSNumber> list);
 }
 
 extension InstantiatedExtension on Instantiated {
-  external List<Instantiated> fieldList;
-  external List<Instantiated> get getList;
-  external set setList(List<Instantiated> list);
-  external List<Instantiated> returnList();
-  external void consumeList(List<Instantiated> list);
+  external JSList<int> fieldList;
+  external JSList<int> get getList;
+  external set setList(JSList<int> list);
+  external JSList<int> returnList();
+  external void consumeList(JSList<int> list);
 }
 
 extension type InstantiatedExtensionType._(JSObject _) {
   // Test generic types where all the type parameters are instantiated.
-  external InstantiatedExtensionType(List<JSNumber> list);
-  external List<InstantiatedExtensionType> fieldList;
-  external List<InstantiatedExtensionType> get getList;
-  external set setList(List<InstantiatedExtensionType> list);
-  external List<InstantiatedExtensionType> returnList();
-  external void consumeList(List<InstantiatedExtensionType> list);
+  external InstantiatedExtensionType(JSList<int> list);
+  external JSList<int> fieldList;
+  external JSList<int> get getList;
+  external set setList(JSList<int> list);
+  external JSList<int> returnList();
+  external void consumeList(JSList<int> list);
 }
 
 void main() {}