+++ tests/lib/js/static_interop_test/js_array_test.dart
@@ -10,7 +10,7 @@ library js_array_test;
 import 'dart:js_interop';
 
 @JS()
-external dynamic eval(String code);
+external void eval(String code);
 
 // dart:js_interop top-levels do return-type checks so if the call to these
 // getters succeed, it's enough to know they can be interoperable.